some bugs fixed

+++ SparseMatrixCW/SparseMatrix.cpp
@@ -15,6 +15,6 @@ SparseMatrix::SparseMatrix(int n, int elements) {
 SparseMatrix::~SparseMatrix() {}
 
 void SparseMatrix::setMatrix(vector<int> AA, vector<int> JR, vector<int> JC) {}
-void SparseMatrix::setMatrix(int** matrix) {}
+void SparseMatrix::setMatrix(double** matrix) {}
 void SparseMatrix::print() {}
 void SparseMatrix::setFromFile(ifstream& ofs) {}
