+++ SparseMatrixCW/SparseMatrix.h
@@ -17,7 +17,7 @@ public:
 	SparseMatrix(int n, int elements);
 	virtual ~SparseMatrix();
 	virtual void setMatrix(vector<int> AA, vector<int> JR, vector<int> JC);
-	virtual void setMatrix(int** matrix);
+	virtual void setMatrix(double** matrix);
 	virtual void print();
 	virtual void setFromFile(ifstream& ofs);
 };
